Support Weigthed Routing Wizard when workloads are added/removed
